+++ packages/utoo-web/src/hmr/index.ts
@@ -24,5 +24,33 @@
  * and is injected into the preview iframe by the build system.
  */
 
+export {
+  type BaseUpdate,
+  type BuildingAction,
+  type BuiltAction,
+  type CompilationError,
+  type EcmascriptMergedUpdate,
+  type HMR_ACTION_TYPES,
+  HMR_ACTIONS_SENT_TO_BROWSER,
+  type HmrActionType,
+  type HmrClientMessage,
+  type HmrIssue,
+  type HmrIssueSource,
+  type IssuesUpdate,
+  type NotFoundUpdate,
+  type PartialUpdate,
+  type ReloadAction,
+  type ResourceIdentifier,
+  type RestartUpdate,
+  type SyncAction,
+  type TurbopackConnectedAction,
+  type TurbopackMessageAction,
+  type TurbopackSubscribeMessage,
+  type TurbopackUnsubscribeMessage,
+  type TurbopackUpdate,
+  type UpdateEndMessage,
+  type UpdateInfo,
+  type UpdateMessage,
+  type UpdateStartMessage,
+} from "@utoo/pack-shared";
 export { type HmrClient, HmrServer, type HmrServerOptions } from "./HmrServer";
-export * from "./types";
