+++ packages/utoo-web/src/project/Project.ts
@@ -1,7 +1,6 @@
-import { handleIssues } from "@utoo/pack-shared";
+import { handleIssues, type UpdateMessage } from "@utoo/pack-shared";
 import * as comlink from "comlink";
 import { HmrClient, HmrServer } from "../hmr";
-import { UpdateMessage } from "../hmr/types";
 import { Fork, HandShake } from "../message";
 import { installServiceWorker } from "../serviceWorker/install";
 import {
